+++ src/java/dao/QuoteDAO.java
@@ -7,35 +7,35 @@ import java.util.List;
 
 public interface QuoteDAO {
     /** Lấy toàn bộ quote của 1 opportunity, mới nhất trước */
-    List<Quote> getByOpportunityId(long opportunityId);
+    List<Quote> getByOpportunityId(int opportunityId);
 
     /** Lấy 1 quote theo id */
-    Quote getById(long id);
+    Quote getById(int id);
 
     /**
      * Tạo quote mới ở trạng thái Draft.
      * @return id của quote mới tạo
      */
-    long create(long opportunityId, String subject, BigDecimal grandTotal, LocalDate validUntil, long createdBy);
+    int create(int opportunityId, String subject, BigDecimal grandTotal, LocalDate validUntil, int createdBy);
 
     /**
      * Chuyển Draft → Sent.
      * Gọi trước hasActiveSent() để đảm bảo không có quote Sent nào khác.
      */
-    void send(long quoteId);
+    void send(int quoteId);
 
     /**
      * Chuyển Sent → Accepted.
      * Đồng thời set Opportunity.stage = Won (trong 1 transaction).
      */
-    void accept(long quoteId, long opportunityId);
+    void accept(int quoteId, int opportunityId);
 
     /** Chuyển Sent → Rejected */
-    void reject(long quoteId);
+    void reject(int quoteId);
 
     /** Xóa quote (chỉ khi Draft) */
-    void delete(long quoteId);
+    void delete(int quoteId);
 
     /** Kiểm tra xem opportunity đã có quote nào đang Sent chưa */
-    boolean hasActiveSent(long opportunityId);
+    boolean hasActiveSent(int opportunityId);
 }
